server/pyproject.toml
[project]
name = "template-server"
version = "0.1.0"
description = "HTTP boilerplate with hexagonal and DDD structure"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.116.1",
  "uvicorn[standard]==0.35.0",
  "pydantic-settings==2.10.1",
  "email-validator==2.2.0",
  "sqlalchemy==2.0.43",
  "psycopg[binary]==3.2.13",
  "pymysql==1.1.2",
  "pymongo==4.14.1",
  "python-jose[cryptography]==3.5.0",
  "passlib[bcrypt]==1.7.4",
  "bcrypt==4.0.1",
]

[project.optional-dependencies]
dev = [
  "httpx==0.28.1",
  "pytest==8.4.1",
  "ruff==0.12.11",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["api", "contexts", "shared"]

[tool.pytest.ini_options]
pythonpath = ["."]
